+++ moor/lib/src/runtime/query_builder/statements/select/select.dart
@@ -4,10 +4,19 @@ part of '../../query_builder.dart';
 /// a table.
 typedef OrderClauseGenerator<T> = OrderingTerm Function(T tbl);
 
+/// The abstract base class for all select statements in the moor api.
+///
+/// Users are not allowed to extend, implement or mix-in this class.
+@sealed
+abstract class BaseSelectStatement extends Component {
+  int get _returnedColumnCount;
+}
+
 /// A select statement that doesn't use joins
 class SimpleSelectStatement<T extends Table, D extends DataClass>
     extends Query<T, D>
-    with SingleTableQueryMixin<T, D>, LimitContainerMixin<T, D>, Selectable<D> {
+    with SingleTableQueryMixin<T, D>, LimitContainerMixin<T, D>, Selectable<D>
+    implements BaseSelectStatement {
   /// Whether duplicate rows should be eliminated from the result (this is a
   /// `SELECT DISTINCT` statement in sql). Defaults to false.
   final bool distinct;
@@ -21,6 +30,9 @@ class SimpleSelectStatement<T extends Table, D extends DataClass>
   /// The tables this select statement reads from.
   @visibleForOverriding
   Set<TableInfo> get watchedTables => {table};
+
+  @override
+  int get _returnedColumnCount => table.$columns.length;
 
   @override
   void writeStartPart(GenerationContext ctx) {